[HBASE-7358] Adding Compaction write size metric
Author: manukranthk

Summary:
Adding metric to capture the amount of data(in bytes) that was written during a compaction. This along with the volume of data flushed might give better estimates on the usage patterns. (There was some problem with my
git commit messages so, I was not able to commit to the earlier revision(D653175), hence adding a new revision.)

Test Plan: Adding unit tests.

Reviewers: kannan, liyintang

Reviewed By: kannan

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D657388

Task ID: 1944049

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1422319 13f79535-47bb-0310-9956-ffa450edef68