[master] Add support in hbase-daemon for regionserver stop-for-restart
Author: mbm

Summary:
Since hadoopctl wants (and kind of needs) to own the stopping and starting
of its daemons for restarts, this exposes an interface to the the jruby used
in doing the localised restart of the regionserver. hadoopctl can pick this
up and use this during a restart.

Test Plan:
This has been tested successfully on hbase430.frc1 (though it gets reverted
by fbdeploy - or something similar - periodically). Basically - the tests
for the positive situation and negative situations with hadoopctl (other
types of stop, hbase-daemon.sh that doesn't support stop-for-restart etc)
all now pass.

Reviewers: arjen, aleksandr, paultuckfield, rthiessen, kannan, kranganathan, liyintang

Reviewed By: arjen

CC: hbase-prodeng@, hbase-eng@

Differential Revision: https://phabricator.fb.com/D645234

Task ID: 906907

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1424044 13f79535-47bb-0310-9956-ffa450edef68