[HBASE-7276] Keep track of the num of rows being read and updated in RegionServer
Author: liyintang

Summary:
There was some confusions before whether RS shall maintain the num of read/write rpc call ops or the num of rows being read/updated. After some discussions with Adela, we felt like there is useless to maintain the rpc call ops since region server had already kept track of the multiput, next and get ops anyway.

So this diff is to keep track of the num of rows being read and updated in RS. In addition, I have removed the per-region level read/write cnt since it is not cose-effective. (Expensive to maintain it but not much value gained)

Test Plan: going to test it on the dev cluster and shadow cluster

Reviewers: adela

Reviewed By: adela

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D667725

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1426612 13f79535-47bb-0310-9956-ffa450edef68