[HBASE-7491] testBatchedUpload test cases more deterministic
Author: rshroff

Summary:
Instead of randomly killing any region server, the test case now kills the
region server to which the next put operation is going to be performed.

It also fixes a bug in BatchedUpload code path, where it now updates the
map with the new location of the region incase it was moved.

Test Plan: Ran TestBatchedUpload test. MR Unit Test suite in progress

Reviewers: aaiyer, liyintang

Reviewed By: liyintang

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D669544

Task ID: 1981445

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1429042 13f79535-47bb-0310-9956-ffa450edef68