[0.89-fb] [master] Fix the order of handling MSG_REPORT_EXITING
Author: aaiyer

Summary:
Before this fix, we process all the region closes, before marking
the RS as dead. This causes an issue where all the transient region
assignments are placed on the exiting server (if it happens to be the
    primary). This would not happen, if the server were in the list
of dead servers.

But, because we add the server to the list of dead servers, after
processing the region closes, the server seems "alive and active"
during the assignment process. Fixing the order, so that the
region will be assigned to a different RS, (considering the exiting
    RS as dead.)

Test Plan:
test suite. only failure TestHRegionCloseRetry, also fails
without the diff

Push to TSH025 master and make sure that the regions are not assigned
back to the exiting server.

4 /tmp/2013-02-06_14_39_23_hbase-trunk_fixRSExitingOrder-mrFailures  Failed on
MR:  TestHRegionClose TestHRegionCloseRetry TestRegionServerMetrics
TestThriftServer
1 /tmp/2013-02-06_14_39_23_hbase-trunk_fixRSExitingOrder-localFailures  Failed
on Local:  TestHRegionCloseRetry
mv target/surefire-reports
mr_unit/2013-02-06_14_39_23_hbase-trunk_fixRSExitingOrder

Reviewers: liyintang

Reviewed By: liyintang

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D701100

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1459461 13f79535-47bb-0310-9956-ffa450edef68