[HBASE-8111] Avoids un necessary DNS lookup for master web ui page
Author: rshroff

Summary:
Currently, whenever you load the master web ui, it will fetch the
hostname by giving the address to generate the RS url. This is a DNS
lookup, which might be slow sometimes. Hence, when we have a lot of RS like
SH003(349), these DNS timeouts cause the loading of the web ui to fail.

This failure will lead to alarm when the cluster goes into production.

The fix is to use already cached hostName to avoid the DNS lookup.

Test Plan:
tested it on SH003
http://hbaseproxy001.prn1.facebook.com/hbase_master/searchhbase003-frc1/master.jsp

Reviewers: liyintang, aleksandr

Reviewed By: aleksandr

CC: arjen, aleksandr, hbase-eng@

Differential Revision: https://phabricator.fb.com/D739051

Task ID: 2185822

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1459473 13f79535-47bb-0310-9956-ffa450edef68