[HBASE-8113] Bring back the seek+read functionality to Pluto.
Author: manukranthk

Summary: It was observed that the seek + read gives better performance than pread when performing relatively large scans(scan which cannot be cached). Hence reverting [HBASE-7408] Removes the pread/seek+read option and defaults it to pread(D661385). Currently this feature is intended to be used only with the ClientSideScan(D708638). This can be potentially made configurable in the future.

Test Plan: Unit test on MR

Reviewers: liyintang, rshroff

Reviewed By: liyintang

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D736962

Task ID: 2179482

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1459474 13f79535-47bb-0310-9956-ffa450edef68