[HBASE-8206] Thread safety issues with Compression.Algorithm.GZ
Author: manukranthk

Summary: The Compression.Algorithm was not completely thread safe. Hence making it thread safe. Trying to fix this the same way it was fixed in HBase trunk(HBASE-5458).

Test Plan: Tested by running ClientSideScan(which opens store files in parallel to perform a client side merge) on the of the region servers on MORSE002-FRC3 cluster. Will try to write a test case which simulates exactly the same way. Also let me know if there is a better way to test this out.

Reviewers: aaiyer, liyintang, nspiegelberg, kannan

Reviewed By: liyintang

CC: hbase-eng@, sdong

Differential Revision: https://phabricator.fb.com/D752641

Task ID: 2225566

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1462244 13f79535-47bb-0310-9956-ffa450edef68