[0.89-fb] [HBASE-8216] Enable the serverManager to respect the config change for Power failures
Author: aaiyer

Summary:
ServerManager has a bug in which we only re-evaluate information about
the rack failure, after all the servers are back. Need to fix this.

Test Plan:
run MR tests.
Do a rack failure testing on DL and ensure that the region servers
time out correctly, when part of the rack recovers.

Reviewers: liyintang, kannan, kranganathan

Reviewed By: liyintang

Differential Revision: https://phabricator.fb.com/D750252

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1462571 13f79535-47bb-0310-9956-ffa450edef68