[HBASE-8249] Adds client side context information
Author: rshroff

Summary:
The motive is to capture failures at per op level. There might be situations in
which the retries will succeed, but the overall operation failed from AppServer
perspective, as the timeout is only 5 secs. Hence, we need a way to get more
context about these operations and record them in Scuba.

This diff adds more information to client side operations performed via
HTable. It adds a ThreadLocal Object at HConnectionManager#TableServers level to
record the exceptions occuring for the current operation. Provides APIs
to fetch this context information and hooks to enable/disable this
recording of context.

Test Plan: mr, Also modified TestHCM to verify this feature.

Reviewers: aaiyer, liyintang

Reviewed By: liyintang

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D750533

Task ID: 2186415

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1464739 13f79535-47bb-0310-9956-ffa450edef68