[HBASE-8083] For RetriesExhausted Exception, tells whether any of the exception was thrown by RegionServer or not
Author: rshroff

Summary:
Whenever AppServer gets an errors(mostly RetriesExhaustedException),
they invalidate the cache for that user. The reason is that they are not
sure if the operation was performed on the RegionServer or not.

This change adds an API to tell whether any of the exception was thrown
by RegionServer or not. In case its not thrown by RegionServer, then
they will not invalidate the cache.

Test Plan:
mr
Will add a unit test to create different errors.

Reviewers: aaiyer, liyintang

Reviewed By: liyintang

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D761556

Task ID: 2253805

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1470788 13f79535-47bb-0310-9956-ffa450edef68