[0.89-fb] [master] handle rootRegion close correctly
Author: aaiyer

Summary:
Ensure that if a root region closes, the master is
able to handle it cleanely.

there have been situations where the root region
remains unassigned; because the master does not
complete the shutdown because meta regions are
not online.

This is due to a bug in the way ProcessRegionClose
looks to check if the parent table is up. For root,
      there is no parent region, so the availability
      should always return true.

Test Plan: unit tests

Reviewers: rshroff, liyintang, adela

Reviewed By: adela

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D777909

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1477748 13f79535-47bb-0310-9956-ffa450edef68