[0.89-fb] [master] Handle missed region open messages from master -> RS
Author: aaiyer

Summary:
We have seen issues on the new hardware where network issues
may cause the MSG_REGION_OPEN from master to the RS to be lost.

This causes the master to keep waiting for the region to open up.
However, since the message is lost, the RS will never open it.

This causes downtime and involves manual intervention to fix, using
hbck.

The diff will ensure that
 (i) the master will retry sending these messages
on future heartbeats, if the RS does not acknowledge the message.
 (ii) The regionserver is able to handle duplicate region open messages
 by maintaining a set of regionOpening regions; and ignoring duplicate
 requests.

Test Plan:
Changed TestRegionRebalancing to cause message loss and
test.

will run the test suite

Reviewers: liyintang, rshroff, adela

Reviewed By: rshroff

CC: hbase-eng@, manukranthk

Differential Revision: https://phabricator.fb.com/D771498

Task ID: 123, 2191086

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1477749 13f79535-47bb-0310-9956-ffa450edef68