[HBASE-8500] Adds a RollingRestart utility for a region server
Author: rshroff

Summary:
The change adds a basic RollingRestart utility which can be performed at
region server level. It also adds a basic RegionChecker which gets the
availability of the region during the rolling restart.

The changes in the master code path is to have more finer grained
control on region servers which are black listed. Plus a bug fix in the
canOffloadTo()

You can now trigger a RollingRestart of region server on the host by
issuing:
*HBASE/bin/hbase org.apache.hadoop.hbase.util.RollingRestart -c -s hbasedev128.ash3

  // Test

Test Plan:
mr, plus did a RollingRestart of a region server on dev
cluster(Sample output https://phabricator.fb.com/P2126823)

Reviewers: aaiyer, liyintang, manukranthk, adela

Reviewed By: aaiyer

CC: hbase-eng@, paultuckfield

Differential Revision: https://phabricator.fb.com/D796086

Task ID: 2229110

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1480007 13f79535-47bb-0310-9956-ffa450edef68