[HBASE-8524] [0.89-fb] Fix indefinite wait() in HBaseClient
Author: aaiyer

Summary:
HBaseClient threads wait() indefinitely on the Connection to get
a response from the server. However, if the server does not respond
(for whatever reasons) the client could be waiting forever.

Moreover, it will be difficult to guarantee rpcTimeout settings if
we wait forever.

Changing the indefinite wait to a loop of timed waits; Bailing out
if we wait for longer than the rpcTimeout.

Test Plan:
mvn tests
TBD:
try to get this on ods and re-pro with TF

Reviewers: liyintang, rshroff, shaneh

Reviewed By: liyintang

CC: hbase-eng@, mbm, mycnyc

Differential Revision: https://phabricator.fb.com/D806604

Task ID: 2362485

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1483520 13f79535-47bb-0310-9956-ffa450edef68