[HBASE-8544] Added a utility to reload configurations in Region Server
Author: gauravm

Summary: Created a utility to tell the Region Server to reload its configurations from disk. We need to fill in what happens when HRegion::updateConfiguration() and Store::updateConfiguration() are called. Currently it just updates the Configuration object, but we might be interested in updating certain other objects as well.

Test Plan: Unit tests

Reviewers: liyintang, aaiyer, rshroff, manukranthk

Reviewed By: rshroff

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D806711

Task ID: 2258346

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1483527 13f79535-47bb-0310-9956-ffa450edef68