[HBASE-8185] Adding functionality to create ClientLocalScanner without creating hardlinks.
Author: manukranthk

Summary: This feature targets the scenarios where the user will be responsible for creating a full table snapshot and reads from the snapshot. This helps in scenarios where we need to scan from a full snapshot to keep the view of the data consistent over scanner(mapper) failures on the scan job.

Test Plan: Unit tests.

Reviewers: aaiyer, liyintang, rshroff

Reviewed By: liyintang

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D830097

Task ID: 2430669

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1489990 13f79535-47bb-0310-9956-ffa450edef68