[HBASE-8688] [89-fb] Adding a choice of single/multiple throttler for callqueue
Author: adela

Summary: I am adding two throttles such that we will throttle the larger and smaller responses differently. Looking at the charts, in a normal state we are not returning responses bigger than 25MB, so I am adding throttler which will throttle the responses <= 25MB and the size of that throttler would be 256MB (10 threads could be writing in the same time). The other throttler would be 768MB and would be for the larger requests. This would potentially help in a case when the client is requesting large responses: the smaller ones would not be blocked. Note that previously we had only 1 throttle of size 1GB so now we are basically dividing the space between 2 of them.

Test Plan: to be defined

Reviewers: aaiyer, liyintang, rshroff

Reviewed By: rshroff

CC: hbase-eng@, rshroff

Differential Revision: https://phabricator.fb.com/D778501

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1489991 13f79535-47bb-0310-9956-ffa450edef68