[0.89-fb] [master] Allow regionServerReport to run with shorter RPC timeouts.
Author: aaiyer

Summary:
Our RS Expiry/Heartbeat mechanism relies on waiting 20 sec
before declaring a RS to be dead. The rationale behind this
is to give the RS 20 consecutive attempts before considering it
dead.

However, if the RPC timeout is set to 60 sec (by default); it is possible
that one regionserverReport takes up to 60 sec (to fail), and that
single failed call can cause the RS to expire.

We should shorten the RPC timeout for the regionserver report to ensure
we give the RS multiple chances before shooting it down.

2) Also including a bug fix around the instrumentation to log regionserver
reports. If the call to master.regionserverReport is unsuccessful; we should
preserve the old behavior and let the exception out. Specifically, the rest
of the code that follows the rsReport should only be executed, if the report
was successful.

Test Plan: unit tests

Reviewers: rshroff, manukranthk

Reviewed By: manukranthk

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D840170

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1492337 13f79535-47bb-0310-9956-ffa450edef68