[HBASE-8714] Add sync_file_range(WAIT) option to flush and compactions.
Author: pritam

Summary:
The idea here is that we don't want runaway compactions and
flushes to write faster than disk to OS/raid cache and then hit the
disks hard on a writeback. For this purpose we invoke a
sync_file_range(WAIT) every megabyte which forces the regionserver to wait
until the amount of data that it has written hasn't been persisted to
disk.

Test Plan:
1) All unit tests.
2) Simple unit test added to trigger option.

Reviewers: liyintang, rshroff, paultuckfield, aaiyer, sdong

Reviewed By: aaiyer

CC: hbase-eng@, hdfs-dev@

Differential Revision: https://phabricator.fb.com/D839449

Task ID: 2233461

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1492338 13f79535-47bb-0310-9956-ffa450edef68