[HBASE-8185] Fixing 'ClientLocalScanner without creating Hardlinks' to not delete the hardlinks.
Author: manukranthk

Summary: Fixing ClientLocalScanner's 'without hardlinks' case to not delete the hardlinks. Earlier, these hardlinks used to be deleted since they are created by the ClientLocalScanner itself. But in this approach the hardlinks should be assumed to be created externally and so they cannot be deleted.

Test Plan: Unit test.

Reviewers: rshroff, shaneh

Reviewed By: rshroff

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D842232

Task ID: 2483974

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1492339 13f79535-47bb-0310-9956-ffa450edef68