[HBASE-8905] Closes the connections after the split is computed
Author: rshroff

Summary:
Currently in the TableInputFormatBase, after getting the input splits,
it is not closing the connections to the client. Most of the time this
is fine as we don't trigger to many jobs from a particular node.

However, there is per client limit on the number of connections in the
Zookeeper and in case we trigger more jobs from a node, it will run to
errors.

Test Plan:
test on SH004. Started job and checked the logs to make sure,
it closes the connection.

P2303750

Reviewers: liyintang

Reviewed By: liyintang

CC: hbase-eng@, chaoyc

Differential Revision: https://phabricator.fb.com/D873353

Task ID: 2558905

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1504218 13f79535-47bb-0310-9956-ffa450edef68