[HBASE-9102] Fixes TestEncoderSeekers test series
Author: rshroff

Summary:
The hashcode() of the BlockCacheKey was not using the ordinal,
which prevents it from differenting the type of encoding, and
in hbase we have the option of having different encoding scheme in
BlockCache and in the HFile.

The fix adds back the ordinal check.

Test Plan: run the MR unit test

Reviewers: manukranthk, liyintang, adela

Reviewed By: adela

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D1001245

Task ID: 2973454

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1530745 13f79535-47bb-0310-9956-ffa450edef68