[HBASE-9896] Adds mappers/job option for HBaseStreaming
Author: rshroff

Summary:
Currently the split algo only supports mappers/region for any MR
job for HBase Streaming. The reason is to have locality of the
region server hosting the region and the mapper task.
However in certain scenarios, we need to have a strict limit on
the number of mappers per job.

This changes adds an additional parameter to control the total
number of mappers/job. The numMappersPerRegion  and numMappersPerJob
are mutually exclusive.

Test Plan: tested with various startRow/stopRow/num_mapper configs on SH004

Reviewers: manukranthk, liyintang

Reviewed By: liyintang

CC: hbase-eng@, chaoyc, nzhang

Differential Revision: https://phabricator.fb.com/D1032630

Task ID: 2381046

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1539432 13f79535-47bb-0310-9956-ffa450edef68