[HBASE-10042] Updating the libthrift, zookeeper and other jars, and adding a maven rule for transitive dependencies.
Author: gauravm

Summary:
Updating the libthrift, zookeeper, mockito, log4j and slf4j jars, since one of the AppServer team is running into problems with the old jars that HBase pulls in.

Also added the requireUpperBoundDeps as a rule during the building process. This will force us to specify the correct version of a dependency, if it is required transitively by multiple dependencies.

Test Plan: Ran the MR Unit tests, and there are no more failures than trunk.

Reviewers: aaiyer, liyintang, adela, rshroff, manukranthk, fan

Reviewed By: fan

CC: hbase-eng@, ryandm

Differential Revision: https://phabricator.fb.com/D1075110

Task ID: 3165197

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1546436 13f79535-47bb-0310-9956-ffa450edef68