[0.89-fb][HBASE-10109] Balance #regions/RS better when creating new table
Author: adela

Summary:
When a new table is created we place the primary locations for
the regions in a round robin fashion. Because we do this for every newly
created table it is possible that some regionservers end up with more
regions than the others and if there are more tables in the cluster the
difference between a min and max #regions per RS can be larger.
One small optimization is to sort the RS in ascending order before
assigning them regions.

Test Plan: wrote unit test: TestRegionPlacement.testPrimaryPlacement

Reviewers: liyintang, manukranthk, rshroff

Reviewed By: rshroff

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D1090198

Task ID: 3339329

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1550235 13f79535-47bb-0310-9956-ffa450edef68