[HBASE-10236] Avoid creating extra connection when locating a region
Author: gauravm

Summary:
In Swift, I found that inside locateRegionInMeta(), we create a connection to meta, before checking if the region's location is in the cache. Most of the time, we expect a hit, hence most of the time, we end up creating a useless connection object. In swift, a lot of these objects caused the thread pool to get exhausted while starting up.

This might not be a problem in trunk, but still it seems like an improvement.

Test Plan: Unit Tests

Reviewers: liyintang, aaiyer, rshroff, manukranthk, adela, fan

Reviewed By: fan

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D1110985

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1553722 13f79535-47bb-0310-9956-ffa450edef68