[HBASE-10205] Fix ConcurrentModificationException in BucketAllocator
Author: arjen

Summary: The different WriterThreads in the BucketCache share a reference to a single BucketAllocator. There is race condition between BucketAllocator.allocateBlock(), which is called when flushing each cache entry, and BucketAllocator.getIndexStatistics(), which usually gets called by the WriterThread after the queue with entries has been flushed. This diff adds some synchronization around these operations. Locking here is ok as calls to the BucketAllocator happen outside the critical path of adding to the cache. Removal of items from the cache happen by the compaction thread or on close, both should be ok here.

Test Plan: Ran the loadtest on one of dev clusters for 12+ hours. Without this diff exceptions would show up within a few hours.

Reviewers: liyintang, aaiyer, rshroff

Reviewed By: liyintang

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D1108464

Task ID: 3337650

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1553892 13f79535-47bb-0310-9956-ffa450edef68