[HBASE-10342] RowKey Prefix Bloom Filter
Author: liyintang

Summary:
When designing HBase schema for some use cases, it is quite common to combine multiple information within a RowKey. For instance, assuming that RowKey is defined as md5(id1) + id1 + id2, and user wants to scan all the RowKeys which starting by id1. In such case, the RowKey Prefix Bloom Filter is able to cut more unnecessary seeks during the scan.

This diff is to implement the RowKey Prefix Bloom Filter in HBase. And the user could define the length of the RowKey prefix through the HTable schema.

Test Plan: Run through all the unit tests

Reviewers: manukranthk, adela

Reviewed By: manukranthk

CC: hbase-dev@, hbase-eng@, arice, daviddeng

Differential Revision: https://phabricator.fb.com/D1141078

Task ID: 3548019

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1567261 13f79535-47bb-0310-9956-ffa450edef68