[HBASE-10360] Auto initialized all the HRegionLocations even though the forceFetch is false.
Author: liyintang

Summary:
The HTable.getCachedHRegionLocations(false) only returns all the cached HRegionLocations based on the original design. Per discussions with Calypso team, we decided to enhance the API so that even though the forceFetch is set as false, if the HRegionLocations for that table has not been initialized yet, it will fetch all the HRegionLocations from the META table.

This enhancement has been implemented in HConnectionManager instead of HTable layer. Therefore, multiple HTable instances for the same table could share this "initialized" state.

Test Plan: TestHTable

Reviewers: manukranthk

Reviewed By: manukranthk

CC: hbase-dev@

Differential Revision: https://phabricator.fb.com/D1157181

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1567266 13f79535-47bb-0310-9956-ffa450edef68