[HBASE-10437] Adding HTable API function to get histogram information from all the regions of a table.
Author: manukranthk

Summary: Adds a getHistogramForAllRegions to HTable which concurrently acquires the Histogram data from all the regions and returns the results back to the client in the ascending order of the regions. This also adjusts the start and end row of the last buckets to the region boundaries so that its easier for the users while using the feature.

Test Plan: Added unit tests

Reviewers: adela, liyintang, aaiyer, gauravm

Reviewed By: liyintang

CC: hbase-eng@, daviddeng

Differential Revision: https://phabricator.fb.com/D1148039

Task ID: 3588805

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1568011 13f79535-47bb-0310-9956-ffa450edef68