[HBASE-9896] Fix the end row bug in the TableInputFormatBase
Author: rshroff

Summary:
In case the the endrow is not specified, the end row is taken from the
HexSplit Class, which defaults it to 'ffffffff'. This will exclude all
the rows starting with ffffffff and after. The fix honors the endrow
from the config instead of using the one from the HexSplit Class.

Test Plan: MR unit tests, and custom job

Reviewers: manukranthk, liyintang

Reviewed By: manukranthk

CC: hbase-dev@, syyang, chaoyc

Differential Revision: https://phabricator.fb.com/D1171876

Task ID: 3599597

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1568490 13f79535-47bb-0310-9956-ffa450edef68