[HBASE-10450] Fix the Thrift test cases
Author: rshroff

Summary:
The TestThriftServer is the base class for testing all HBase operations
from thrift via different thrift servers. Hence a failure in TestThriftServer
fails multiple test cases. Also TestThriftServer takes the Hbase.Iface as
an input to perform all the operations. The Hbase.Iface can either be the standard
HBase thrift client or the ThriftServerRunner.HBaseHandler (which acts as an
intermediate between the client and the hbase client). The bug is that
the TestThriftServer now assumes the Hbase.Iface to be always of type HBaseHandler,
which causes the problem. Hence, removing it.

Test Plan: run unit tests.

Reviewers: manukranthk, daviddeng

Reviewed By: daviddeng

CC: hbase-eng@

Differential Revision: https://phabricator.fb.com/D1186452

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1572205 13f79535-47bb-0310-9956-ffa450edef68