[HBASE-10777] Restart Embedded Thrift Server in case of failures
Author: rshroff

Summary:
When the overall load on the embedded thrift server increases, the
TThreadedThriftServer crashes. The RegionServer is still alive and
serving regions but clients talking to embedded thrift server cannot
talk with those regions.

This is a hack for now to restart the embedded thrift server in case
of such failures. TThreadedThriftServer is outdated and we are planning to
get rid of the proxy thrift server moving forward. Hence, I have added
this hack.

Test Plan: existing MR unit tests. Test it on couple of RS on SH003

Reviewers: manukranthk

Reviewed By: manukranthk

CC: hbase-dev@, csliu, chaoyc, dcapra

Differential Revision: https://phabricator.fb.com/D1222171

Task ID: 3797478

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1584163 13f79535-47bb-0310-9956-ffa450edef68