[master] First commit of Endpoint implementation.
Author: daviddeng

Summary:
An endpoint is a method running on the server. Some aggregation work can be done on the server and only final results are returned back to the client.

Current endpoints have the following restrictions, may be removed later:

An example of aggregation is shown in `TestEndpoint`

Test Plan: `TestEndpoint`

Reviewers: adela, manukranthk, aaiyer, fan, liyintang, gauravm

Reviewed By: adela

CC: hbase-eng@, andrewcox

Differential Revision: https://phabricator.fb.com/D1223428

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1584179 13f79535-47bb-0310-9956-ffa450edef68