[master] Fix unstable testcase TestGlobalMemStoreSize
Author: daviddeng

Summary:
The reason of being unstable is that after you force a flush on regions of META table, it's possible some other updates to META table is inserted before asserting. Then the mem-store is non empty again and the testcase failed.

The workaround is to skip assertion for the server with META table's regions.

More information is logged.

Test Plan: `TestGlobalMemStoreSize`

Reviewers: liyintang, manukranthk, gauravm, elliott

Reviewed By: elliott

CC: hbase-eng@, elliott

Differential Revision: https://phabricator.fb.com/D1275557

Task ID: 4138742

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1588113 13f79535-47bb-0310-9956-ffa450edef68